feat: add Document structure to tree
This is a simple version of the final document and it already looks
pretty cool. BTW, I managed to finally compile the document in Windows.
It was a major pain to set up.

The following failed:
    - In MSYS: `pacman -Sy mingw-w64-x86_64-texlive-core
    - In UCRT64: `pacman -Sy mingw-w64-ucrt-x86_64-texlive-core`
in building the `fmtutil` and `updmap` due to an issue with my ZLib
installation [header=(1.3) vs dll=(1.3.1)]

VERDICT: Lesson learnt
Assigned-to: @m-ikiara